+++ src/main/java/org/mju_likelion/festival/booth/service/BoothQueryService.java
@@ -27,10 +27,10 @@ import org.mju_likelion.festival.booth.dto.response.BoothQrResponse;
 import org.mju_likelion.festival.booth.dto.response.SimpleBoothResponse;
 import org.mju_likelion.festival.booth.dto.response.SimpleBoothResponses;
 import org.mju_likelion.festival.booth.util.qr.manager.BoothQrManager;
-import org.mju_likelion.festival.common.circuit_breaker.FallBackUtil;
 import org.mju_likelion.festival.common.exception.BadRequestException;
 import org.mju_likelion.festival.common.exception.ForbiddenException;
 import org.mju_likelion.festival.common.exception.NotFoundException;
+import org.mju_likelion.festival.common.util.circuit_breaker.FallBackUtil;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
