+++ src/main/java/org/mju_likelion/festival/common/util/annotaion/page_size/PageSize.java
@@ -1,4 +1,4 @@
-package org.mju_likelion.festival.common.annotaion.page_size;
+package org.mju_likelion.festival.common.util.annotaion.page_size;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
