+++ src/main/java/org/mju_likelion/festival/lost_item/controller/LostItemController.java
@@ -11,10 +11,10 @@ import static org.mju_likelion.festival.common.api.ApiPaths.SEARCH_LOST_ITEMS;
 import jakarta.validation.Valid;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
-import org.mju_likelion.festival.common.annotaion.page_number.PageNumber;
-import org.mju_likelion.festival.common.annotaion.page_size.PageSize;
 import org.mju_likelion.festival.common.authentication.AuthenticationPrincipal;
-import org.mju_likelion.festival.common.enums.SortOrder;
+import org.mju_likelion.festival.common.util.annotaion.page_number.PageNumber;
+import org.mju_likelion.festival.common.util.annotaion.page_size.PageSize;
+import org.mju_likelion.festival.common.util.enums.SortOrder;
 import org.mju_likelion.festival.lost_item.dto.request.CreateLostItemRequest;
 import org.mju_likelion.festival.lost_item.dto.request.LostItemFoundRequest;
 import org.mju_likelion.festival.lost_item.dto.request.UpdateLostItemRequest;
