+++ src/main/java/org/mju_likelion/festival/lost_item/service/LostItemQueryService.java
@@ -8,9 +8,9 @@ import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
-import org.mju_likelion.festival.common.circuit_breaker.FallBackUtil;
-import org.mju_likelion.festival.common.enums.SortOrder;
 import org.mju_likelion.festival.common.exception.NotFoundException;
+import org.mju_likelion.festival.common.util.circuit_breaker.FallBackUtil;
+import org.mju_likelion.festival.common.util.enums.SortOrder;
 import org.mju_likelion.festival.lost_item.domain.LostItem;
 import org.mju_likelion.festival.lost_item.domain.LostItemDetail;
 import org.mju_likelion.festival.lost_item.domain.SimpleLostItem;
