+++ src/main/java/org/mju_likelion/festival/term/service/TermQueryService.java
@@ -9,9 +9,9 @@ import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import org.mju_likelion.festival.common.circuit_breaker.FallBackUtil;
 import org.mju_likelion.festival.common.exception.BadRequestException;
 import org.mju_likelion.festival.common.exception.type.ErrorType;
+import org.mju_likelion.festival.common.util.circuit_breaker.FallBackUtil;
 import org.mju_likelion.festival.term.domain.Term;
 import org.mju_likelion.festival.term.domain.repository.TermJpaRepository;
 import org.mju_likelion.festival.term.domain.repository.TermQueryRepository;
